requirements.txt
torch==1.13.1+cu117
torchvision==0.14.1+cu117
torchaudio==0.13.1
--extra-index-url https://download.pytorch.org/whl/cu117
flask
pillow
numpy<2
pandas
matplotlib
gunicorn